Splitting server/start.js into two files
