+++ tests/server/index.spec.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const app = require('../../server');
+const chai = require('chai');
+const supertest = require('supertest');
+
+const expect = chai.expect;
+
+describe('Express App', () => {
+  it('should run without throwing errors', () => {
+
+  });
+});
